requirements.txt
attrs==19.3.0
Automat==20.2.0
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
docopt==0.6.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Script==2.0.6
Flask-Twisted==0.1.2
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
mysql-connector-python==8.0.21
observable==1.0.3
packaging==20.4
pathtools==0.1.2
pluggy==0.13.1
protobuf==3.12.4
py==1.9.0
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.1
pytest-watch==4.2.0
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
Twisted==20.3.0
urllib3==1.25.10
watchdog==0.10.3
Werkzeug==1.0.1
zipp==3.1.0
zope.interface==5.1.0
